requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
gunicorn==21.2.0
motor==3.3.1
pydantic==2.4.2
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
google-generativeai==0.3.1
python-multipart==0.0.6
nltk==3.8.1
rake-nltk==1.0.6
sentence-transformers==2.2.2
httpx==0.25.1
websockets==11.0.3 